TST: Fix failing test

+++ pandas/tests/groupby/test_libgroupby.py
@@ -287,7 +287,7 @@ def test_cython_group_mean_not_datetimelike_but_has_NaT_values():
 def test_cython_group_mean_Inf_at_begining_and_end():
     # GH 50367
     actual = np.array([[np.nan, np.nan], [np.nan, np.nan]], dtype="float64")
-    counts = np.array([0, 0])
+    counts = np.array([0, 0], dtype="int64")
     data = np.array(
         [[np.inf, 1.0], [1.0, 2.0], [2.0, 3.0], [3.0, 4.0], [4.0, 5.0], [5, np.inf]],
         dtype="float64",
